Add timeout option to query methods
Problem

Large reads from watchman can return EAGAIN even if the socket is set to
blocking as watchman collates the large list of changes.  As libwatchman
provides no ability to re-read from the socket, this prevents further
progress as re-issuing the query results in the same EAGAIN response.

Solution

Call select() during queries to emulate blocking on available I/O.  This
is an API change, so I added it as a new method,
watchman_do_query_timeout to maintain backwards compatibility.