configure.ac: Comment indicating why you might get a AM_PATH_CHECK() syntax error
If you run the configure script and get an error about check.h, then
re-run configure after installing it you'll get:

    ./configure: line 11641: syntax error near unexpected token `0.9.4'
    ./configure: line 11641: `AM_PATH_CHECK(0.9.4)'

That's because the AM_PATH_CHECK macro comes with check itself, so we
need to re-run autogen.sh. Add a comment about this in case anyone else
runs into the same issue.